Change tree trimming method
previously: first ancestor of sequence X with occurence of mutation Y.

But this had the problem of the trim being bad if that one sequence is placed outside of the clade we are aiming for (well, it fails because the trimmed tree contains too few sequences)

Now: find all nodes with occurrence of mutations, which do have mutations in 'required_mutations' and do not have mutations in 'forbidden mutations. Then check there is only one such clade with > min_tips descendants, and trim to that clade.